Test: Test new fetch from Gerrit ability
A new feature has been added to the gerrit checkout action that is
supposed to try checking out the change directly from Gerrit if it fails
to get it from the GitHub mirror. This puts a job in place in an
advisory position to allow us to validate over time if this is working
correctly.

Issue-ID: CIMAN-33
Change-Id: Idabf008cc4e6f29190714cd606e4475936640c2f
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>